pipeline/requirements.txt
google-cloud-storage
google-cloud-bigquery
tensorflow
numpy